change3.B_device&timer&protocol & .M_oled&ds18b20
device
    修改命名规范
    修改设备的结构体定义
    增加构造函数，并在构造时进行判空处理
    优化头文件结构
timer
    在头文件中引出定时器相关宏定义
    将延时参数统一修改为uint64_t
    删除宏函数，统一使用内联
protocol
    修改命名
oled&ds18b20
    随device修改初始化结构
    将oled的字库修改为utf8-BOM编码，此编码下使用汉字不会出现warning
    因为编码修改，将oled的printchinese函数及中文数组函数的长度由2增加至3

+++ protocol.c
@@ -10,6 +10,6 @@ __weak bool DEVCMNI_SDA_OWRE_In(void) { return HIGH; }
 __weak bool DEVCMNI_SDO_In(void) { return HIGH; }
 __weak void DEVCMNI_CS_Out(bool pot) {}
 __weak void DEVCMNI_Error(int8_t err) {}
-__weak void DEVCMNI_Delayus(uint16_t us) {}
-__weak void DEVCMNI_Delayms(uint16_t ms) {}
-__weak int8_t DEVCMNI_Delayus_paral(uint16_t us) { return 1; }
+__weak void DEVCMNI_Delayus(uint64_t us) {}
+__weak void DEVCMNI_Delayms(uint64_t ms) {}
+__weak int8_t DEVCMNI_Delayus_paral(uint64_t us) { return 1; }
